Update patch set 4
Patch Set 4:

Ok, this is working as I wanted now - the review triggers the umbrella job, which in turn executes the docker build and then CSIT in sequence, and finally the failing CSIT votes the review down. The question now is how do we move forward with this? I'll need to have this in master in order to test my next changes to JJB templates in sandbox.

Patch-set: 4