Update patch set 4
Patch Set 4:

There doesn't seem to be any reaction from project owners (neither to this commit nor to the onap-discuss mail that Morgan sent) so maybe let's merge this and move on? It will have the effect of failing to verify any commits to music for a short while but I should fix that quite quickly in the following changes.

Patch-set: 4
Reviewer: Gerrit User 3949 <3949@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>